requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
uritemplate==4.1.1
